==> notebooks/02_merge_and_feature_reduction.ipynb <==
import pandas as pd

%cd /content/drive/MyDrive/Agriculture App/agriculture-predictor-planner

pd.set_option('display.max_columns', None)
pd.set_option('display.max_rows', None)


**Merging the soil/weather dataset with crop dataset**

# Load .xls file
soil_weather_data = pd.read_excel('data/cleaned/cleaned_soil_weather_data.xlsx')

# Load .csv file
crop_data = pd.read_csv('data/cleaned/cleaned_crop_produce_data.csv')



print("Soil and Weather Dataset Shape:", soil_weather_data.shape)
print("Crop Dataset Shape:", crop_data.shape)

# Display column names
print("Soil and Weather Dataset Columns:", soil_weather_data.columns)
print("Crop Dataset Columns:", crop_data.columns)

# Check for common keys
print("Common Keys:", set(soil_weather_data.columns).intersection(crop_data.columns))



print(soil_weather_data.shape)
print(crop_data.shape)

#Dropping the non useful redundant columns
columns_to_drop = crop_data.columns[[5,6,7,20,21,22,32,33,34,41,42,43,57,58,59,68,69,70]]
crop_data = crop_data.drop(columns=columns_to_drop, axis=1)
print(crop_data.columns)


#Checking the shape and common columns after column drops
print(crop_data.shape)
print("Common Keys:", set(soil_weather_data.columns).intersection(crop_data.columns))

# Merge on common columns
merged_data = pd.merge(soil_weather_data, crop_data, on=[ 'Year', 'State Code', 'Dist Code'], how='inner')

# Preview merged data
print(merged_data.info())
print(merged_data.head())
print(merged_data.shape)


print(list(merged_data.columns))

merged_data.info(verbose=True)

#Dropping uneccesary columns
merged_data.rename(columns={'State Name_x': 'State Name', 'Dist Name_x': 'Dist Name'}, inplace=True)
merged_data = merged_data.drop(merged_data.columns[[106, 107]], axis=1)
print(merged_data.info(verbose=True))

#Converting data types so that district code, year, and state code are not treated as numeric during calculations
merged_data = merged_data.astype({'Dist Code':'category', 'Year':'category', 'State Code':'category', 'State Name':'category', 'Dist Name':'category'})

print(merged_data.info(verbose=True))

#Saving the merged file
merged_path = 'data/merged/crop_soil_weather_merged.csv'
merged_data.to_csv(merged_path, index=False)
print(f"Merged crop produce data saved to {merged_path}")
